+++ R/cache-helpers.R
@@ -833,11 +833,18 @@ copyFile <- Vectorize(copySingleFile, vectorize.args = c("from", "to"))
     otherFns <- character()
   }
 
-  # Figure out if it is in a .parseModule call, if yes, then extract the module
-  doEventFrameNum <- .grepSysCalls(scalls, "\\.parseModule")
-  #doEventFrameNum <- which(startsWith(as.character(scalls), prefix = ".parseModule"))
+  # Figure out if it is in SpaDES.core call, which could be either a .parseModule or .runModuleInputObjects call...
+  #   , if yes, then extract the module
+  doEventFrameNum <- .grepSysCalls(scalls, "\\.parseModule|\\.runModuleInputObjects")
   if (length(doEventFrameNum)) {
-    module <- get("m", envir = sys.frame(doEventFrameNum[2])) # always 2
+    module <- get0("m", envir = sys.frame(tail(doEventFrameNum, 1)))
+    if (is.null(module)) { # this block should cover any other cases, though is likely unnecessary
+      # This whole mechanism is predicated on the module name being called "m" in the above 2 functions
+      moduleEnv <- whereInStack("m")
+      module <- get0("m", envir = moduleEnv)
+    }
+
+    # module <- get("m", envir = sys.frame(doEventFrameNum[2])) # always 2
     otherFns <- c(paste0("module:", module), otherFns)
   }
   unique(otherFns)